monitoring: Do not delete namespace in use
Delete 'monitoring' namespace alongside when destroying prometheus or
grafana only if it becomes empty.

+++ grafana/destroy-grafana.sh
@@ -7,5 +7,9 @@ source "$SCGC_HOME/common.sh"
 
 pkill kubectl
 helm uninstall -n monitoring grafana
-kubectl delete namespace monitoring
+
+PODS=$(kubectl get pods -n monitoring --no-headers 2>/dev/null | grep -v grafana | wc -l)
+if [[ "$PODS" == 0 ]]; then
+	kubectl delete namespace monitoring
+fi
