+++ prometheus/destroy-prometheus.sh
@@ -7,5 +7,9 @@ source "$SCGC_HOME/common.sh"
 
 fuser -k 9090/tcp
 helm uninstall -n monitoring prometheus
-kubectl delete namespace monitoring
+
+PODS=$(kubectl get pods -n monitoring --no-headers 2>/dev/null | grep -v prometheus | wc -l)
+if [[ "$PODS" == 0 ]]; then
+	kubectl delete namespace monitoring
+fi
